Fix a signed/unsigned mismatch compiler warning on VS2008
git-svn-id: svn://scm.orgis.org/mpg123/trunk@2312 35dc7657-300d-0410-a2e5-dc2837fedb53

+++ src/libmpg123/id3.c
@@ -274,7 +274,7 @@ char *next_text(char* prev, int encoding, size_t limit)
 
 		text += width;
 	}
-	if(text-prev >= limit) text = NULL;
+	if((size_t)(text-prev) >= limit) text = NULL;
 
 	return text;
 }
